pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "portal-use"
version = "0.1.0"
description = "Wayland-native desktop computer use MCP server via XDG Desktop Portal"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0",
    "dbus-next>=0.2.3",
    "Pillow>=10.0",
    # PyGObject (gi) must be installed via system package: python3-gi
    # gstreamer1.0-pipewire must be installed via system package
]

# NOTE: portal-use = "server:main_sync" — main_sync wraps asyncio.run(main())
# and will be added to server.py before this entry point is usable.
[project.scripts]
portal-use = "server:main_sync"

[tool.hatch.build.targets.wheel]
packages = ["portal"]
include = ["server.py"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
